test-requirements.txt
hacking
sphinx
nose
openstack.nose_plugin
coverage
mock
pylint
python-openstackclient
boto
python-swiftclient
